requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.5
djangorestframework==3.14.0
flake8==6.1.0
flake8-bugbear==23.9.16
flake8-class-attributes-order==0.1.3
flake8-clean-block==0.1.2
flake8-commas==2.1.0
flake8-indent-in-def==0.1.2
flake8-length==0.3.1
flake8-multiline-containers==0.0.19
flake8-newspaper-style==0.4.3
flake8-plugin-utils==1.3.3
flake8-quotes==3.3.2
flake8-return==1.2.0
isort==5.12.0
mccabe==0.7.0
pep8-naming==0.13.3
psycopg2==2.9.8
pycodestyle==2.11.0
pyflakes==3.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
